+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1
+
+* fix: depend on correct `IconTheme`
+* refactor: rename `FeatherIconsTheme` to `FeatherIconTheme`
+
 ## 0.2.0
 
 * fix: remove SVGs for assets because they are generated as a SVG strings